+++ src/interfaces/wasi/mod.rs
@@ -1,0 +1,6 @@
+mod wit_generation;
+#[cfg(feature = "wasi_impl")]
+mod wit_implementation;
+
+#[cfg(test)]
+mod tests;